app/requirements.txt
flask
flask-restx
fastapi
uvicorn[standard]
llama-cpp-python
huggingface-hub
pydantic
